Non-bootstraped build must employ cmake() generators
When building non-boostrapped cmake but with disabled some features,
cmake-data won't get into build root and cmake(FindPkg) and
cmake(UsePkg) Provides will be missing from just-built cmake-data
package. (This bug is not visible with full-fledged build because
cmake-data with the RPM dependency generator is installed
transitively).

This patch ensures that cmake-data is in the build root when not
bootstrapping.